all midrounds have same chance
