+++ Content.Server/Nyanotrasen/Events/Components/MidRoundAntagSpawnLocationComponent.cs
@@ -1,0 +1,8 @@
+namespace Content.Server.StationEvents
+{
+    [RegisterComponent]
+    public sealed class MidRoundAntagSpawnLocationComponent : Component
+    {
+
+    }
+}
