--- Content.Server/Nyanotrasen/Fugitive/FugitiveSpawnLocationComponent.cs
@@ -1,8 +1,0 @@
-namespace Content.Server.Fugitive
-{
-    [RegisterComponent]
-    public sealed class FugitiveSpawnLocationComponent : Component
-    {
-
-    }
-}
