+++ Resources/Maps/angle.yml
@@ -78488,7 +78488,7 @@ entities:
   - color: '#FF0000FF'
     type: AtmosPipeColor
 - uid: 9550
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 31.5,2.5
     parent: 9
@@ -81965,7 +81965,7 @@ entities:
     parent: 9
     type: Transform
 - uid: 9955
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 74.5,-30.5
     parent: 9
@@ -124578,67 +124578,67 @@ entities:
     type: Transform
   - type: ActiveEmergencyLight
 - uid: 16005
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 34.5,-20.5
     parent: 9
     type: Transform
 - uid: 16006
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 35.5,-21.5
     parent: 9
     type: Transform
 - uid: 16007
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 57.5,5.5
     parent: 9
     type: Transform
 - uid: 16008
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 47.5,6.5
     parent: 9
     type: Transform
 - uid: 16009
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 55.5,3.5
     parent: 9
     type: Transform
 - uid: 16010
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 48.5,4.5
     parent: 9
     type: Transform
 - uid: 16011
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 17.5,11.5
     parent: 9
     type: Transform
 - uid: 16012
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -6.5,-11.5
     parent: 9
     type: Transform
 - uid: 16013
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -21.5,-62.5
     parent: 9
     type: Transform
 - uid: 16014
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 19.5,-50.5
     parent: 9
     type: Transform
 - uid: 16015
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 55.5,9.5
     parent: 9
@@ -128991,7 +128991,7 @@ entities:
     parent: 9
     type: Transform
 - uid: 16698
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 47.5,12.5
     parent: 9
@@ -130879,13 +130879,13 @@ entities:
     parent: 9
     type: Transform
 - uid: 16983
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 1.5,-56.5
     parent: 9
     type: Transform
 - uid: 16984
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 0.5,-71.5
     parent: 9
@@ -136450,13 +136450,13 @@ entities:
     parent: 9
     type: Transform
 - uid: 17809
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -11.5,-1.5
     parent: 9
     type: Transform
 - uid: 17810
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -15.5,-5.5
     parent: 9
