+++ Resources/Maps/arena.yml
@@ -40537,7 +40537,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 558
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -19.5,-29.5
     parent: 0
@@ -40984,7 +40984,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 626
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -3.5,-23.5
     parent: 0
@@ -118162,7 +118162,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 12451
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -45.5,-35.5
     parent: 0
@@ -121817,7 +121817,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 12987
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -45.5,-40.5
     parent: 0
@@ -193276,13 +193276,13 @@ entities:
     parent: 0
     type: Transform
 - uid: 22500
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -44.5,-51.5
     parent: 0
     type: Transform
 - uid: 22501
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -5.5,-62.5
     parent: 0
@@ -193307,13 +193307,13 @@ entities:
   - color: '#0335FCFF'
     type: AtmosPipeColor
 - uid: 22504
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 8.5,-52.5
     parent: 0
     type: Transform
 - uid: 22505
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 35.5,-55.5
     parent: 0
@@ -193480,19 +193480,19 @@ entities:
     parent: 0
     type: Transform
 - uid: 22532
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 60.5,-13.5
     parent: 0
     type: Transform
 - uid: 22533
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 49.5,-16.5
     parent: 0
     type: Transform
 - uid: 22534
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -31.5,17.5
     parent: 0
@@ -193872,7 +193872,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 22594
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -31.5,-23.5
     parent: 0
@@ -193884,7 +193884,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 22596
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -33.5,-41.5
     parent: 0
@@ -193897,7 +193897,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 22598
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 7.5,-20.5
     parent: 0
@@ -193916,7 +193916,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 22601
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 57.5,26.5
     parent: 0
@@ -197645,7 +197645,7 @@ entities:
     parent: 0
     type: Transform
 - uid: 23217
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 11.5,-21.5
     parent: 0
