+++ Resources/Maps/lighthouse.yml
@@ -29357,7 +29357,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 497
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 8.5,-70.5
     parent: 100
@@ -49189,7 +49189,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 3501
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -28.5,43.5
     parent: 100
@@ -57708,7 +57708,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 4812
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -31.5,-7.5
     parent: 100
@@ -143383,7 +143383,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 17057
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -38.5,25.5
     parent: 100
@@ -143419,7 +143419,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 17063
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -50.5,51.5
     parent: 100
@@ -143457,19 +143457,19 @@ entities:
     parent: 100
     type: Transform
 - uid: 17069
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 25.5,23.5
     parent: 100
     type: Transform
 - uid: 17070
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 40.5,-5.5
     parent: 100
     type: Transform
 - uid: 17071
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 0.5,-3.5
     parent: 100
@@ -143483,7 +143483,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 17073
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 0.5,-9.5
     parent: 100
@@ -143525,19 +143525,19 @@ entities:
     parent: 100
     type: Transform
 - uid: 17078
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 1.5,-28.5
     parent: 100
     type: Transform
 - uid: 17079
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 18.5,-9.5
     parent: 100
     type: Transform
 - uid: 17080
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -0.5,13.5
     parent: 100
@@ -143828,7 +143828,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 17123
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -17.5,27.5
     parent: 100
@@ -146296,7 +146296,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 17298
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -5.5,63.5
     parent: 100
@@ -158516,7 +158516,7 @@ entities:
     parent: 100
     type: Transform
 - uid: 18986
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 35.5,39.5
     parent: 100
