+++ Resources/Maps/ovni.yml
@@ -73586,55 +73586,55 @@ entities:
   - color: '#0000FFFF'
     type: AtmosPipeColor
 - uid: 9332
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 13.5,11.5
     parent: 8
     type: Transform
 - uid: 9333
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 12.5,27.5
     parent: 8
     type: Transform
 - uid: 9334
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -8.5,15.5
     parent: 8
     type: Transform
 - uid: 9335
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -0.5,-1.5
     parent: 8
     type: Transform
 - uid: 9336
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -18.5,23.5
     parent: 8
     type: Transform
 - uid: 9337
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -11.5,31.5
     parent: 8
     type: Transform
 - uid: 9338
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 28.5,-2.5
     parent: 8
     type: Transform
 - uid: 9339
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 30.5,-19.5
     parent: 8
     type: Transform
 - uid: 9340
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 10.5,-26.5
     parent: 8
