+++ Resources/Maps/shoukou.yml
@@ -16620,7 +16620,7 @@ entities:
     parent: 34
     type: Transform
 - uid: 1072
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -41.5,-9.5
     parent: 34
@@ -78750,7 +78750,7 @@ entities:
     parent: 34
     type: Transform
 - uid: 10235
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 24.5,8.5
     parent: 34
@@ -83333,13 +83333,13 @@ entities:
     - 10341
     type: DeviceList
 - uid: 10833
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -8.5,-16.5
     parent: 34
     type: Transform
 - uid: 10834
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -23.5,-5.5
     parent: 34
@@ -83351,7 +83351,7 @@ entities:
     parent: 34
     type: Transform
 - uid: 10836
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -28.5,-14.5
     parent: 34
@@ -85911,19 +85911,19 @@ entities:
     parent: 34
     type: Transform
 - uid: 11239
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: 9.5,-14.5
     parent: 34
     type: Transform
 - uid: 11240
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -10.5,-41.5
     parent: 34
     type: Transform
 - uid: 11241
-  type: SpawnPointLocationFugitive
+  type: SpawnPointLocationMidRoundAntag
   components:
   - pos: -53.5,-4.5
     parent: 34
